pyproject.toml
[project]
name = "ahp-2-0"
version = "0.1.0"
description = "AI Health Passport - Enterprise Grade Backend"
authors = [{name = "Antigravity", email = "ai@nirixa.com"}]
readme = "README.md"
requires-python = "^3.11"

dependencies = [
    "fastapi[all]>=0.110.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "redis>=5.0.0",
    "slowapi>=0.1.9",
    "structlog>=24.1.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "psycopg2-binary>=2.9.9",
    "opentelemetry-api>=1.23.0",
    "opentelemetry-sdk>=1.23.0",
    "opentelemetry-instrumentation-fastapi>=0.44b0",
    "arq>=0.25.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "Pillow>=10.2.0",
    "pytesseract>=0.3.10",
    "cryptography>=42.0.0",
    "twilio (>=9.10.3,<10.0.0)",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "bandit>=1.7.7",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
